feat(sharing): add Supabase auth, nutrition spaces, and collaborative sync
Introduce email login and granular nutrition sharing so families can create spaces, invite members and groups, and keep grocery lists and meal plans in sync per space via Supabase.

+++ composeApp/src/commonMain/kotlin/app/mymultiverse/kmp/domain/nutrition/WeekCalendar.kt
@@ -46,7 +46,7 @@ object WeekCalendar {
     ): Int? {
         if (weekKeyFor(today) != weekKey) return null
         val start = LocalDate.parse(weekKey)
-        val index = today.toEpochDays() - start.toEpochDays()
+        val index = (today.toEpochDays() - start.toEpochDays()).toInt()
         return index.takeIf { it in 0 until WeeklyMealPlan.DAYS_IN_WEEK }
     }
